requirements.txt
colorama
Flask
h5py
keras
numpy
Pillow
tensorflow